requirements.txt
python-dotenv~=0.14.0
PyMySQL~=0.10.1
boto3~=1.15.18